Refactor app, link, and folder desktop generation scripts into a single script [endlessm/eos-shell#444]
