--- desktop/csv_to_desktop_links.py
@@ -1,141 +1,0 @@
-#!/usr/bin/env python
-
-import os
-import errno
-import re
-
-class DesktopWriter:
-
-    def __init__(self, csv_path, desktop_dir):
-        self._csv_path = csv_path
-        self._desktop_dir = desktop_dir
-        self.make_sure_path_exists(desktop_dir)
-
-    def make_sure_path_exists(self, path):
-        try:
-            os.makedirs(path)
-        except OSError as exception:
-            if exception.errno != errno.EEXIST:
-                raise
-
-    def locale_string(self, locale):
-        if locale == 'default':
-            return ''
-        else:
-            return '[' + locale + ']'
-
-    def url_to_exec(self, url):
-        exec_string = 'epiphany-browser %s' % url
-        return exec_string
-
-    def _write_key(self, desktop_file, fields, key):
-        # Write a line for each localized version of the key
-        for locale in self._locales:
-            field = fields[self._indexes[key][locale]]
-            # Other than the default, omit blank values
-            if field or locale == 'default':
-                line = '%s%s=%s\n' % (key, self.locale_string(locale), field)
-                desktop_file.write(line)
-
-    def _write_desktop_file(self, fields):
-        desktop_id = fields[0]
-        desktop_path = os.path.join(self._desktop_dir,
-                                    'eos-link-' + desktop_id +
-                                    '.desktop')
-        desktop_file = open(desktop_path, 'w')
-        desktop_file.write('[Desktop Entry]\n')
-        desktop_file.write('Version=1.0\n')
-        self._write_key(desktop_file, fields, 'Name')
-        self._write_key(desktop_file, fields, 'Comment')
-	desktop_file.write('Type=Application\n')
-
-        have_localized_url = False
-        for locale in self._locales:
-            index = self._indexes['URL'][locale]
-            localized_url = fields[index]
-            if localized_url and locale != 'default':
-                have_localized_url = True
-                break
-
-        if have_localized_url:
-            exec_string = 'eos-exec-localized'
-            index = self._indexes['URL']['default']
-            url = fields[index]
-            exec_string += " '" + self.url_to_exec(url) + "'"
-            for locale in self._locales:
-                index = self._indexes['URL'][locale]
-                localized_url = fields[index]
-                if localized_url and locale != 'default':
-                    localized_exec = self.url_to_exec(localized_url)
-                    exec_string += " " + locale + ":'" + localized_exec + "'"
-        else:
-            index = self._indexes['URL']['default']
-            url = fields[index]
-            exec_string = self.url_to_exec(url)
-
- 	desktop_file.write('Exec=%s\n' % exec_string)
-        self._write_key(desktop_file, fields, 'Icon')
-
-    def _add_index(self, key, locale, index):
-        self._locales.add(locale)
-
-        try:
-            inner_dict = self._indexes[key]
-        except:
-            inner_dict = {}
-        
-        inner_dict[locale] = index
-        self._indexes[key] = inner_dict
-
-    def _parse_header(self, header):
-        # Note: for now, the Desktop and AppStore columns are ignored
-        # They will be used later to specify what is available by default
-        # on the desktop and in the app store for each user personality
-
-        # Set of all locales
-        self._locales = set()
-
-        # Dictionary that relates keys and locales to indexes
-        self._indexes = {}
-
-        # Find all the locales specified in the header
-        locale_keys = ['Name', 'Comment', 'URL', 'Icon']
-        index = 0
-
-        # For each field in the header
-        fields = header.split(',')
-        for field in fields:
-            for key in locale_keys:
-                if field.startswith(key):
-                    if field == key:
-                        # Non-localized field
-                        locale = 'default'
-                    else:
-                        # Localized field
-                        regex = '^' + key + '\[(.+)\]$'
-                        match_result = re.match(regex, field)
-                        if match_result:
-                            locale = match_result.group(1)
-                        else:
-                            print 'Invalid localized field header:', field
-                            exit(1)
-                    self._add_index(key, locale, index)
-            index += 1
-
-    def write_desktop_files(self):
-        csv_file = open(self._csv_path, 'r')
-
-        # Parse the first line header
-        header = csv_file.readline().rstrip()
-        self._parse_header(header)
-
-        # For each remaining line after the header
-        for line in csv_file:
-            fields = line.rstrip().split(',')
-            self._write_desktop_file(fields)
-    
-        csv_file.close()
-
-if __name__ == '__main__':
-    desktop_writer = DesktopWriter('links.csv', 'links')
-    desktop_writer.write_desktop_files()
